Add --amount flag and bump to 1.5.0

+++ src/index.ts
@@ -1,11 +1,11 @@
 #!/usr/bin/env node
 
-import { ChildProcessByStdio, spawn } from "child_process";
+import * as assert from "assert";
+import { spawn } from "child_process";
 import { program } from "commander";
-import { Readable } from "stream";
 
 program
-  .version("1.0.0")
+  .version("1.5.0")
   .name("smell-test")
   .description("Runs a cli command and waits on a specific log line to appear")
   .argument("<log-line>", "The log line to wait for")
@@ -14,12 +14,23 @@ program
     "The command to run. If you want to pass flags, use --"
   )
   .option("--quiet", "Don't output anything")
+  .option(
+    "-a, --amount <amount>",
+    "The amount of times to check for the log line",
+    "1"
+  )
   .action(
     async (
       logLine: string,
       command: string[],
-      options: { quiet?: boolean }
+      options: { quiet?: boolean; amount: string }
     ) => {
+      const amount = parseInt(options.amount);
+      assert(!isNaN(amount), "Amount must be a number");
+      if (!options.quiet) {
+        console.log(`Waiting for log line: ${logLine}`);
+        console.log(`Running command: ${command.join(" ")}`);
+      }
       const child = spawn(command[0], command.slice(1), {
         stdio: ["inherit", "pipe", "pipe"],
       });
@@ -27,8 +38,9 @@ program
         child.stderr.pipe(process.stderr);
         child.stdout.pipe(process.stdout);
       }
-      child.stdout.on("data", (data) => tryFind(data, logLine, child));
-      child.stderr.on("data", (data) => tryFind(data, logLine, child));
+      let count = 0;
+      child.stdout.on("data", (data) => tryFind(data, logLine));
+      child.stderr.on("data", (data) => tryFind(data, logLine));
       child.on("close", () => {
         if (!options.quiet) {
           console.error(
@@ -39,19 +51,20 @@ program
         }
         process.exit(1);
       });
+      function tryFind(data: Buffer, logLine: string) {
+        const dataString = data.toString();
+        if (dataString.includes(logLine)) {
+          if (!options.quiet) {
+            console.log(`Found log line: ${logLine}`);
+          }
+          if (count >= amount) {
+            child.kill("SIGTERM");
+            process.exit(0);
+          }
+          count++;
+        }
+      }
     }
   );
 
 program.parse(process.argv);
-function tryFind(
-  data: Buffer,
-  logLine: string,
-  child: ChildProcessByStdio<null, Readable, Readable>
-) {
-  const dataString = data.toString();
-  if (dataString.includes(logLine)) {
-    console.log(logLine);
-    child.kill("SIGTERM");
-    process.exit(0);
-  }
-}
